+++ src/main/java/com/rental/models/Bike.java
@@ -1,7 +1,7 @@
 package com.rental.models;
 
 public class Bike extends Vehicle{
-    public Bike(String vehicleId, int price, VehicleTypeEnum vehicleTypeEnum) {
+    public Bike(String vehicleId, double price, VehicleTypeEnum vehicleTypeEnum) {
         super(vehicleId, price, vehicleTypeEnum);
     }
 }
