+++ src/main/java/com/rental/models/Car.java
@@ -1,7 +1,7 @@
 package com.rental.models;
 
 public class Car extends Vehicle{
-    public Car(String vehicleId, int price, VehicleTypeEnum vehicleTypeEnum) {
+    public Car(String vehicleId, double price, VehicleTypeEnum vehicleTypeEnum) {
         super(vehicleId, price, vehicleTypeEnum);
     }
 }
