+++ src/main/java/com/rental/models/Van.java
@@ -1,7 +1,7 @@
 package com.rental.models;
 
 public class Van extends Vehicle{
-    public Van(String vehicleId, int price, VehicleTypeEnum vehicleTypeEnum) {
+    public Van(String vehicleId, double price, VehicleTypeEnum vehicleTypeEnum) {
         super(vehicleId, price, vehicleTypeEnum);
     }
 }
